Check Ai version for proper functionality
Added a new const `versionedCommands` that is an object containing commands or entire subsets of commands that need to be versioned checked for presenting the in the palette. If the minVersion is greater than the current Ai version the command or set of commands will be deleted.

Please note the listed minVersion for each is not correct right now for testing purposes.